requirements.txt
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
plotly>=5.0.0
streamlit>=1.0.0
matlabengine>=9.13.0
pathlib>=1.0.1 